backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1

# Existing Classification Dependencies
playwright>=1.40.0
openai>=1.10.0
beautifulsoup4>=4.12.0
langdetect>=1.0.9
pyyaml>=6.0
pandas>=2.0.0
tqdm>=4.66.0
aiohttp>=3.9.0
lxml>=4.9.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Additional API Dependencies
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyjwt>=2.8.0
